requirements.txt
django==2.0.0
django-redis==4.10.0
django_crontab==0.7.1
paramiko==2.4.2
django-widget-tweaks==1.4.3
pymysql==0.9.3
dnspython=1.16.0
requests==2.21.0
idna==2.8
urllib3==1.24.1
certifi==2019.3.9
chardet==3.0.4
